Fix repeated L1/AUX parsing and require packaged Windows RF-to-TS validation
Keep current and next-frame scheduling separate, allocate repeated PLP/AUX storage, preserve all 48 AUX bits and reject truncated layouts. Add ASan/UBSan field vectors for multiple PLPs, repetition and owned snapshots.

Run the GUI receiver with packaged Windows DLLs on the independent GNU Radio fixture, requiring exact TS before publishing the portable archive.

The supplied recording still fails all 624 BCH words. A separately tested upstream receiver also yields zero TS from it; this remains a diagnostic candidate, not a claim of restored hardware reception.

+++ src/DVB_T2/async_pipeline_payload.h
@@ -63,8 +63,8 @@ private:
     std::vector<l1_postsignalling_aux> aux;
     std::vector<dynamic_plp> dynPlp;
     std::vector<dynamic_plp> dynNextPlp;
-    std::vector<int> dynAux;
-    std::vector<int> dynNextAux;
+    std::vector<uint64_t> dynAux;
+    std::vector<uint64_t> dynNextAux;
 };
 
 // RAII slot for a bounded queued stage. If a queued metacall is discarded
